complete implement user Management
