+++ src/components/Courses/ListCourses/ListCourses.tsx
@@ -1,3 +1,6 @@
+import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
+import EditIcon from '@mui/icons-material/Edit';
+import ExitToAppIcon from '@mui/icons-material/ExitToApp';
 import FilterAltIcon from '@mui/icons-material/FilterAlt';
 import ReplyIcon from '@mui/icons-material/Reply';
 import SaveAltIcon from '@mui/icons-material/SaveAlt';
@@ -9,6 +12,8 @@ import {
   TableContainer,
   TextField
 } from '@mui/material';
+import Divider from '@mui/material/Divider';
+import MenuItem from '@mui/material/MenuItem';
 import Paper from '@mui/material/Paper';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
@@ -17,12 +22,12 @@ import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import { Link as RouterLink } from 'react-router-dom';
 import { courses } from '../../../fakeData/courses';
-import CustomizedMenus from './CourseMenuActions';
+import { CustomizedMenus } from './CourseMenuActions';
 import './ListCourses.css';
 
 const ListCoursesRender = () => {
   // console.log(ListCourses);
-
+  const handleDelete = () => {};
   return (
     <>
       <Container className="main-container">
@@ -88,10 +93,39 @@ const ListCoursesRender = () => {
                       {course.timeUpdate}
                     </TableCell>
                     <TableCell sx={{ paddingRight: '28px' }} align="right">
-                      <CustomizedMenus
-                        linkTo={`/courses/${course.id}/edit`}
-                        courseDetial={`/courses/${course.id}`}
-                      />
+                      <CustomizedMenus>
+                        <Link
+                          component={RouterLink}
+                          to={`/courses/${course.id}/edit`}
+                          className="option-link">
+                          <MenuItem disableRipple>
+                            <EditIcon />
+                            Edit
+                          </MenuItem>
+                        </Link>
+
+                        <Link
+                          component={RouterLink}
+                          to={'#'}
+                          onClick={handleDelete}
+                          className="option-link">
+                          <MenuItem disableRipple>
+                            <DeleteForeverIcon />
+                            Delete
+                          </MenuItem>
+                        </Link>
+                        <Divider sx={{ my: 0.5 }} />
+
+                        <Link
+                          component={RouterLink}
+                          to={`courses/${course.id}`}
+                          className="option-link">
+                          <MenuItem disableRipple>
+                            <ExitToAppIcon />
+                            Go to course
+                          </MenuItem>
+                        </Link>
+                      </CustomizedMenus>
                     </TableCell>
                   </TableRow>
                 ))}
